Create README.md
In this approach I have used the analog reading from the microcontroller input that is connected to the battery and calibrated the sensor reading and utilized the PWM values given to the motors to train a machine learning model. A test code is used to generate the necessary data values to train the machine learning model.